+++ doorstop/core/tree.py
@@ -11,7 +11,6 @@ from itertools import chain
 
 from doorstop.common import DoorstopError, DoorstopWarning, DoorstopInfo
 from doorstop.core.document import Document
-from doorstop.core.item import split_id
 from doorstop.core import vcs
 
 
@@ -201,9 +200,16 @@ class Tree(object):
 
         @raise DoorstopError: if the item cannot be removed
         """
-        item = self.find_item(identifier)
-        item.delete()
-        return item
+        for document in self:
+            try:
+                document.find_item(identifier)
+            except DoorstopError:
+                pass  # item not found in that document
+            else:
+                item = document.remove(identifier)
+                return item
+
+        raise DoorstopError("no matching ID: {}".format(identifier))
 
     def link(self, cid, pid):
         """Add a new link between two items by IDs.
@@ -260,11 +266,13 @@ class Tree(object):
         # Open item
         if launch:
             _open(item.path, tool=tool)
+            # TODO: force the item to reload after editing
+            item._loaded = False  # pylint: disable=W0212
         # Return the item
         return item
 
     def find_document(self, prefix):
-        """Get a document from its prefix.
+        """Get a document by its prefix.
 
         @param prefix: document's prefix
 
@@ -279,7 +287,7 @@ class Tree(object):
         raise DoorstopError("no matching prefix: {}".format(prefix))
 
     def find_item(self, identifier, _kind=''):
-        """Get an item from its ID.
+        """Get an item by its ID.
 
         @param identifier: item ID
 
@@ -287,21 +295,12 @@ class Tree(object):
 
         @raise DoorstopError: if the item cannot be found
         """
-        # Type of item ('parent', 'child', or '') for logging messages
-        _kind = (' ' + _kind) if _kind else _kind
-
-        # Search using the prefix and number
-        prefix = split_id(identifier)[0]
+        _kind = (' ' + _kind) if _kind else _kind  # for logging messages
         for document in self:
-            if document.prefix.lower() == prefix.lower():
+            try:
                 return document.find_item(identifier, _kind=_kind)
-        logging.debug("no matching{} prefix: {}".format(_kind, prefix))
-
-        # Fall back to a search using the exact ID
-        for document in self:
-            for item in document:
-                if item.id.lower() == identifier.lower():
-                    return item
+            except DoorstopError:
+                pass  # item not found in that document
 
         raise DoorstopError("no matching{} ID: {}".format(_kind, identifier))